Move TypeSupport
