+++ tested/evaluators/value.py
@@ -14,8 +14,8 @@ from tested.datatypes import (
 )
 from tested.dodona import ExtendedMessage, Permission, Status, StatusMessage
 from tested.evaluators.common import EvaluationResult, EvaluatorConfig
+from tested.features import TypeSupport
 from tested.internationalization import get_i18n_string
-from tested.languages.config import TypeSupport
 from tested.languages.generation import generate_statement
 from tested.serialisation import (
     ObjectKeyValuePair,
